refactor: Recruit 파일구조 수정 (#239)
