+++ frontend/src/components/Recruit/BProfileSelect/index.jsx
@@ -1,10 +1,9 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { PVImg } from "../../Common";
 
-import * as S from "../style";
+import Presenter from "./presenter";
 
-PSelectBProfile.propTypes = {
+BProfileSelect.propTypes = {
   bplist: PropTypes.arrayOf(
     PropTypes.shape({
       id: PropTypes.number.isRequired,
@@ -50,7 +49,7 @@ PSelectBProfile.propTypes = {
   setRecruit: PropTypes.func.isRequired,
 };
 
-export default function PSelectBProfile({
+export default function BProfileSelect({
   movePage,
   bpList,
   recruit,
@@ -68,31 +67,5 @@ export default function PSelectBProfile({
     movePage(1);
   };
 
-  return (
-    <S.Page>
-      <S.BProfileContainer>
-        <p>모집하기에서</p>
-        <p>어떤 프로필을 사용할 것인가요?</p>
-        <p>사용하실 비즈니스 프로필을 선택하세요</p>
-        <S.BProfileListBox>
-          {bpList.map((a) => {
-            return (
-              <div key={a.id}>
-                <S.BProfileImgBox
-                  onClick={selectBP}
-                  id={a.id}
-                  className={recruit.businessProfile.id === a.id && "selected"}
-                >
-                  <PVImg
-                    img={a.avatar ? a.avatar : "/기본비즈니스프로필.jpeg"}
-                  />
-                </S.BProfileImgBox>
-                <p>{a.name}</p>
-              </div>
-            );
-          })}
-        </S.BProfileListBox>
-      </S.BProfileContainer>
-    </S.Page>
-  );
+  return <Presenter bpList={bpList} selectBP={selectBP} recruit={recruit} />;
 }
