+++ frontend/src/pages/Recruit/index.jsx
@@ -1,12 +1,12 @@
 import React from "react";
-import { CreateFeedContainer } from "../../components/Recruit";
+import RecruitComponent from "../../components/Recruit";
 import { NavBar } from "../../components/Common";
 
 export default function Recruit() {
   return (
     <>
       <NavBar />
-      <CreateFeedContainer />
+      <RecruitComponent />
     </>
   );
 }
